Add shared watchlist storage
