+++ app/api/watchlist/[id]/route.ts
@@ -1,9 +1,7 @@
 export const dynamic = "force-dynamic";
 
 import { NextRequest, NextResponse } from 'next/server';
-
-// External storage reference (shared with main route)
-let watchlists: any[] = [];
+import { watchlists, WatchlistItem } from '../store';
 
 export async function GET(
   request: NextRequest,
@@ -11,7 +9,7 @@ export async function GET(
 ) {
   try {
     const watchlist = watchlists.find(w => w.id === params.id);
-    
+
     if (!watchlist) {
       return NextResponse.json(
         { error: 'Watchlist not found' },
@@ -19,7 +17,7 @@ export async function GET(
       );
     }
 
-    return NextResponse.json({ data: watchlist });
+    return NextResponse.json(watchlist);
   } catch (error) {
     console.error('Error fetching watchlist:', error);
     return NextResponse.json(
@@ -50,6 +48,57 @@ export async function DELETE(
     console.error('Error deleting watchlist:', error);
     return NextResponse.json(
       { error: 'Failed to delete watchlist' },
+      { status: 500 }
+    );
+  }
+}
+
+export async function POST(
+  request: NextRequest,
+  { params }: { params: { id: string } }
+) {
+  try {
+    const body = await request.json();
+    const { ticker } = body;
+
+    if (!ticker) {
+      return NextResponse.json(
+        { error: 'Ticker is required' },
+        { status: 400 }
+      );
+    }
+
+    const watchlist = watchlists.find(w => w.id === params.id);
+
+    if (!watchlist) {
+      return NextResponse.json(
+        { error: 'Watchlist not found' },
+        { status: 404 }
+      );
+    }
+
+    // Check if ticker already exists
+    const existingItem = watchlist.items.find((item: any) => item.ticker === ticker);
+    if (existingItem) {
+      return NextResponse.json(
+        { error: 'Ticker already in watchlist' },
+        { status: 400 }
+      );
+    }
+
+    const newItem: WatchlistItem = {
+      id: Date.now().toString(),
+      ticker: ticker.toUpperCase(),
+      createdAt: new Date().toISOString(),
+    };
+
+    watchlist.items.push(newItem);
+
+    return NextResponse.json(newItem, { status: 201 });
+  } catch (error) {
+    console.error('Error adding watchlist item:', error);
+    return NextResponse.json(
+      { error: 'Failed to add item' },
       { status: 500 }
     );
   }
